requirements.txt
openai
duckduckgo-search 
phidata
uvicorn
sqlalchemy
pgvector
psycopg[binary]
pypdf